BUGFIX: types in message serialization
The osc library changed the osc float serialization to
float64, rest of the code already uses typecasts, so this
single change is enough for now.

Will need to do a cleanup pass later.

+++ v4/clock/messages.go
@@ -39,7 +39,7 @@ type MediaMessage struct {
 	seconds   int32
 	frames    int32
 	remaining int32
-	progress  float32
+	progress  float64
 	paused    bool
 	looping   bool
 	timeStamp osc.Timetag
